Rename from controllers dir to controller
