+++ controller/auth.go
@@ -1,4 +1,4 @@
-package controllers
+package controller
 
 import (
 	"context"
